fastapi_app/requirements.txt
fastapi
uvicorn[standard]
paho-mqtt
confluent-kafka
